refactor: move entidade Tecnico para legacy e ajusta estrutura para uso unificado com Usuario
