+++ src/main/java/br/com/filipecode/DeskhelpApi/chamado/dto/AtualizarChamadoDTO.java
@@ -1,6 +1,4 @@
 package br.com.filipecode.DeskhelpApi.chamado.dto;
 
-import java.util.UUID;
-
-public record AtualizarChamadoDTO(String status, String prioridade, UUID tecnicoID) {
+public record AtualizarChamadoDTO(String status, String prioridade) {
 }
