+++ src/main/java/br/com/filipecode/DeskhelpApi/chamado/entity/Chamado.java
@@ -2,7 +2,6 @@ package br.com.filipecode.DeskhelpApi.chamado.entity;
 
 import br.com.filipecode.DeskhelpApi.chamado.enums.Prioridade;
 import br.com.filipecode.DeskhelpApi.chamado.enums.Status;
-import br.com.filipecode.DeskhelpApi.tecnico.entity.Tecnico;
 import br.com.filipecode.DeskhelpApi.usuario.entity.Usuario;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -47,8 +46,4 @@ public class Chamado {
     @ManyToOne
     @JoinColumn(name = "usuario_id", nullable = false)
     private Usuario usuario;
-
-    @ManyToOne
-    @JoinColumn(name = "tecnico_id")
-    private Tecnico tecnico;
 }
